# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/servers/series": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Servers Usage Series",
        "description": "Return request totals per server over a time window.",
        "operationId": "servers_usage_series_api_servers_series_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bucket"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSeriesDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{name}/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Server Usage",
        "description": "Return request volumes for a server over a time window.",
        "operationId": "server_usage_api_servers__name__usage_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{name}/usage/series": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Server Usage Series",
        "description": "Return bucketed request volumes for a server over a time window.",
        "operationId": "server_usage_series_api_servers__name__usage_series_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bucket"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemCount": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "count"
        ],
        "title": "ItemCount",
        "description": "One labeled usage count."
      },
...
      "SeriesBucket": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Tools"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Methods"
          },
          "clients": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Clients"
          },
          "statuses": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Statuses"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "total",
          "tools",
          "methods",
          "clients",
          "statuses"
        ],
        "title": "SeriesBucket",
        "description": "One time bucket in a usage series."
      },
      "SeriesPoint": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "total"
        ],
        "title": "SeriesPoint",
        "description": "One time bucket's total request count for a server."
      },
      "SeriesReport": {
        "properties": {
          "server": {
            "type": "string",
            "title": "Server"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SeriesBucket"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "server",
          "start",
          "end",
          "bucket",
          "points"
        ],
        "title": "SeriesReport",
        "description": "Request volumes for one MCP server bucketed over a time window."
      },
...
      "ServerSeries": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "name",
          "points"
        ],
        "title": "ServerSeries",
        "description": "Totals-only request series for one server."
      },
...
      "UsageReport": {
        "properties": {
          "server": {
            "type": "string",
            "title": "Server"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Tools"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Methods"
          },
          "clients": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Clients"
          },
          "statuses": {
            "items": {
              "$ref": "#/components/schemas/ItemCount"
            },
            "type": "array",
            "title": "Statuses"
          }
        },
        "type": "object",
        "required": [
          "server",
          "start",
          "end",
          "total",
          "tools",
          "methods",
          "clients",
          "statuses"
        ],
        "title": "UsageReport",
        "description": "Request volumes for one MCP server over a time window."
      },
      "UsageSeriesDocument": {
        "properties": {
          "servers": {
            "items": {
              "$ref": "#/components/schemas/ServerSeries"
            },
            "type": "array",
            "title": "Servers"
          }
        },
        "type": "object",
        "required": [
          "servers"
        ],
        "title": "UsageSeriesDocument",
        "description": "Request totals per mounted server over a time window."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}